:recycle: #19 処理の流れを VideoChatController として分離
